# Performance: register atomic waker lazily in AsyncRingBuf