@agerasev 2023-07-30
30% is really nice performance improvement! It's interesting that registering an `AtomicWaker` is so expensive.

I'm merging the request.

Also I've [added your benchmark](https://github.com/agerasev/ringbuf/commit/03ac5ab55c664b3d92a07f67113a21c48d92da47) to the crate if you don't mind.  

Thank you for your interest in this project and for such an effort to make it better!